Add one-to-one relationship between words and steps
Add one to one relationship between words and steps in their respective
models.
My next steps are going to be to generate migrations with foriegn key
references for each specific relationship, checking for server errors
as I go.
If this goes well, I will begin working through the process of seeding the
database.
If it goes poorly, I will redact the relationship between words and steps,
add a fixed 'index' to sentences, and use that index number in steps
to reference the desired next sentence.