Add curl script to create words
Add curl script to create words.
Adjust words controller to permit specification of sentence.
Test for success.
Confirm that serialization of games and sentences works as expected.
Next step is to check the serialization of steps with a curl script.
If that goes well, I will tinker with the serialization of games.