Add curl script to create steps
Add curl script to create steps
Adjust steps controller to permit word specification.
Next step is to fiddle with the serialization of games to allow
it to display an array of sentences.
Then get sentences to display an array of words.
Then words to display their respective steps.
Then, fiddle with how I want steps to serialize their sentences.