Adjust step and word serializers
Adjust word serializer to show associated step, and step to include the
id of the sentence it references.
Tinkering with update script for a sentence, to see if I can get a step
to refer to a sentence different than its 'parent'.
If that goes well, will beginning to think through seeding DB.