Add curl scripts to update and create steps
Add curl script to update steps, which required adjusting the steps
controller to permit the columns specified in the most recent
migration.
Confirmed that curl script to add steps now works with the specified
columns.
I want to make sure that steps are serializing with a sentence
that they 'own', not merely the sentence that owns the word that owns
them.